fix: refactor user longpoll v3 wrapper
