+++ longpoll-user/v3/errors.go
@@ -1,4 +1,4 @@
-package wrapper
+package v3
 
 import "errors"
