+++ longpoll-user/v3/objects.go
@@ -1,4 +1,4 @@
-package wrapper
+package v3
 
 import (
 	"time"
